+++ sherly-upr-model/src/main/java/com/guzi/upr/model/admin/RoleMenu.java
@@ -18,13 +18,13 @@ public class RoleMenu extends BaseModel {
     /** id */
     @ApiModelProperty("id")
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /** 角色id */
     @ApiModelProperty("角色id")
-    private Integer roleId;
+    private Long roleId;
 
     /** 菜单id */
     @ApiModelProperty("菜单id")
-    private Integer menuId;
+    private Long menuId;
 }
